requirements.txt
numpy>=1.20.0
torch>=1.9.0
matplotlib>=3.5.0
pyyaml>=6.0
tqdm>=4.62.0
scikit-learn>=1.0.0
h5py>=3.6.0
pandas>=1.3.0
networkx>=2.6.0
scipy>=1.7.0
seaborn>=0.11.0
torchvision>=0.10.0
einops>=0.4.1
sympy>=1.9.0
pytest>=6.2.5 